Removed unfunctional test from test_apps.py
This might not be the best method to handle a test, that is simply not
working, but there is currently no working pattern on how to execute the
ready()-method again to test its code.
Please note: While this is not covered in unittests, manual testing will show,
if the code is running successfully, so i.e. django-admin check will raise
the according errors.

This issue is on my TODO list.