Created URL-config (urls.py)
The basic URLs for login/logout are configured.
Please note, that the syntax for including namespace'd sub-url-confs slightly
changed, so regard this for documentation.

The AppConfig() now tries to set default values for login-related settings,
i.e. LOGIN_URL. Django already has default values for this, they (should/have
to) be modified by the user. Include this into documentation aswell.