Integrated with Travis-CI & coveralls.io
Actually, this commit is a fake. The provided .travis.yml was developed and
tested in a seperate branch and is re-created here.

The other branch was a mess, with many one-line-edits to test things out.
Because every minimal change had to be pushed to Github to trigger the Travis
build, it looked really ugly.

Anyway, this commit will be pushed to Github aswell, to confirm the Travis-
configuration, before merging back into development.